+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 0.2.0
+
+* Changed cookbook name to chef_ohai_hints from chef-ohai-hints
+* Added chefspec tests
+* Added serverspec tests
+* Added Rakefile
+* Added .kitchen.ec2.yml for testing on ec2
+* Rubocop and updated metadata.rb, Vagrantfile, Gemfile, Berksfile
+* Updated README.md
+* Updated default recipe to create the necessary files and paths and run ohai
+* Updated awsext.rb to use EC2::Instance and get the tags
+
 # 0.1.0
 
-Initial release of chef-ohai-hints
+* Initial release of chef_ohai_hints